removed deprecated cast char* to const char*
git-svn-id: http://svn.sourceforge.jp/svnroot/konoha/trunk/konoha@1634 413261d9-973d-0410-9b72-d5665a44c00b

+++ src/main/context.c
@@ -432,7 +432,7 @@ static void knh_share_free(Ctx *ctx, knh_share_t *share)
 	share->ClassTable = NULL;
 	DBG_ASSERT(share->ArenaSet != NULL);
 
-	for(i = 0; i < (int)(share->ArenaSetSize); i++) {
+	for(i = 0; i < share->ArenaSetSize; i++) {
 		knh_ArenaSet_t *t = share->ArenaSet + i;
 		KNH_VFREE(ctx, t->head, t->arena_size);
 	}
